Update for AutoDamager

+++ ClickerHeroes/Entities/Hero.cs
@@ -13,7 +13,7 @@ namespace ClickerHeroes.Entities
         public string Description { get; set; }
         public int Level { get; set; }
         public int Price { get; set; }
-        public int Damage { get; set; }
+        public double Damage { get; set; }
         public string ImagePath { get; set; }
         public bool Bought { get; set; }
     }
